api/requirement.txt
tensorflow==2.20.0
tensorflow-serving-api==2.19.1
fastapi==0.116.1
uvicorn==0.35.0
python-multipart==0.0.20
pillow==11.0.0
matplotlib
numpy
